Added a gitignore for the `designs` folder (#33)
This will make it so the user is not presented with files that they should not keep in their git repo. It is common to delete the `designs` folder when trying to clear an error - so I've kept it in the top-level .gitignore file. 

I've added additional space at the end so that when `slm` adds its ignore - it doesn't get confused with the data cache ignores.